fix: add back progress bar
